Move default column class, change consent status classes.
